pyproject.toml
[project]
name = "embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo>=0.18.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
